requirements.txt
pdfminer.six
upstash-redis
xxhash